Dev vinicius

+++ src/Core/Business/ProdutoBusiness.cs
@@ -4,7 +4,7 @@ using Core.Interfaces.Services;
 
 namespace Core.Business;
 
-public class ProdutoBusiness(IProductRepository productRepository) : IProductService
+public class ProdutoBusiness(IProductRepository productRepository, IOrderItensRepository orderItensRepository) : IProductService
 {
     public Produto AddNewProduct(Produto product)
     {
@@ -29,20 +29,30 @@ public class ProdutoBusiness(IProductRepository productRepository) : IProductSer
         return productRepository.GetProducts();
     }
 
-    public Produto? UpdateProduct(Produto product)
+    public Produto UpdateProduct(Produto product)
     {
         product.Validade();
-        var productExist = GetProductById(product.Id);
-        if (productExist != null)
+        Produto productExist = GetProductById(product.Id);
+        if (productExist == null)
         {
-            return productRepository.InsertUpdateProduct(product);
+            throw new ArgumentException("Produto não encontrado");
         }
-
-        return null;
+        productExist.Descricao = product.Descricao;
+        productExist.Preco = product.Preco;
+        productExist.Estoque = product.Estoque;
+        productExist.Categoria = product.Categoria;
+        return productRepository.InsertUpdateProduct(productExist);
     }
 
-    public void DeleteProduct(Produto product)
+    public void DeleteProduct(int id)
     {
+        Produto product = productRepository.GetProductById(id);
+        if (product == null)
+            throw new ArgumentException("Produto não encontrado");
+
+        if (orderItensRepository.CheckProducIsUsedByOrder(id))
+            throw new ArgumentException("Esse produto não pode ser excluido pois é usado em um pedido");
+
         productRepository.DeleteProduct(product);
     }
 }
